Figured out Delayed_Job Gem
So Delayed_Job is a gem that allows us to run a process in the
background. In the case of our app, it would be populating the shows
database with complete values so it takes less time, therefore reducing
timeouts.

However, you need a "worker" to do this, which Heroku charges for. The
easy way out is to do "heroku rake jobs:work" and your machine will act
as worker, but as soon as you exit jobs:work or close your terminal, it
will stop doing it.

Solution? Either deal with timeouts, pay for heroku, or have a PC
running rake jobs:work during our presentation.