Added search functionality, read full commit msg
So the search performs a pretty basic SQL query using the helpers in
rails. It looks very simple but took a few hours to make it work.

Basically if it finds a match for the show, it redirects to that shows
page instead of showing a search results page. If there is no exact
match, then it shows an error reflecting this.

Changes:
Added new routes to routes.rb
Added a search link to application.html.erb
Added a search function to shows_controller.rb
New files: /app/views/shows search.html.erb and noresult.html.erb

If you have ideas to make it better, please let me know or by all means
make some changes.